Reduce QTR PIO program size
I moved the initialization of the OSR, Y, and X registers from the
start of the PIO program and put it into the C code which already
existed to reset the program counter back to the beginning of the
program to start each sensor read. This reduced the size of the PIO
code by 3 to a total of 17 instructions.