Simplify LineSensors API
Previously the LineSensors class was taking a reference to a
LineSensorReadings object to be filled in by the read*() methods in the
class. Those methods still exist but I have added overloads which take
no output parameter and the programmer is instead expected to read
the public rawSensorReadings and/or calibratedSensorReadings member
fields after making a read*() method call to get the latest result.