Start switching to Pololu's 3pi+ 2040 C Library
This is the beginning of my switch to Pololu's 3pi+ 2040 C library as
the lowest level in this Arduino library. This change ports the
following modules:
* ledYellow
* readBatteryMillivolts
* ButtonA, ButtonB, ButtonC
* Motors

A snapshot of the Pololou C library is copied into the
src/pololu-3pi-2040-robot/ folder of this repository. This folder
includes the corresponding LICENSE files from the Pololu repository
which hosts these files. I have that repository as a submodule under
the pio folder but it was difficult to get that submodule to build in
the Arduino environment without moving some files around and applying
patches to the sources. The pio/Makefile is able take the files
from the submodule, copy them into the snapshot location, and apply the
patches. The main reason for the patches are:
* The Arduino RP2040 BSP doesn't include the pico/stdlib.h header
  so any headers that are pulled in by this higher level one need to
  be pulled in explicitly instead.
* gpio_init() and spi_init() also need to have an '_' prefix added to
  account for the name clash with the mbed library.